requirements.txt
asgiref==3.11.1
Django==4.2.16
pillow==11.3.0
psycopg2-binary==2.9.12
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
